Refactor to make the package installable via PyPI
A small refactor has been performed to include the setup.py.
Dependencies are now managed by this file, as well as package
versioning and so on. Internal imports are now imported using
`metafinder.utils.<whatever>` syntax. For clarity, the launcher
(`cli.py`) has been separated from the `processing(...)` method.
This approach is closer to the use of the package as a library
making it closer to be imported as `import metafinder.core`.

Some minor issues have also been addressed:
- The number of the results returned by Google was always 1 less
- A new argument has been provided to set the output folder
- The `results` folder is now created if it does not exist
- `pathlib` package has been used for the creation of the folder
- The usage execution has been clarified
- Versioning has been started as `0.1.0`. This may be changed.

Installation instructions in the `README.md` have been updated.
 Note that these instructions won't be ready until the package
is uploaded to PyPI.